+++ tests/HelperStringTest.php
@@ -374,23 +374,23 @@ class HelperStringTest extends HelperTestCase
         $this->assertNotTrue(HelperString::containsAnyOf('abc', 'd'));
     }
 
-    /**
-     * Test: HelperString::pregMatchAllWithOffsets
-     */
-    public function testPregMatchAllWithOffsets(): void
-    {
-        $this->assertEquals(
-            json_encode([2 => 'a', 10 => 'a', 21 => 'a', 25 => 'a', 30 => 'a', 35 => 'a', 40 => 'a']),
-            json_encode(HelperString::preg_match_all_with_offsets('/a/', 'Fraticinidas velum, tanquam flavum adiurator.')));
-
-        $this->assertEquals(
-            json_encode([11 => 's9', 14 => 'v9', 42 => 'm9']),
-            json_encode(HelperString::preg_match_all_with_offsets('/[a-z]9/', 'Fraticinidas9 v9 el-1 um2, tanquam 9 flavum9 adiurator39.')));
-
-        $this->assertEquals(
-            json_encode([1 => 'raticinidas9', 14 => 'v9', 37 => 'flavum9']),
-            json_encode(HelperString::preg_match_all_with_offsets('/[a-z]+9/', 'Fraticinidas9 v9 el-1 um2, tanquam 9 flavum9 adiurator39.')));
-    }
+//    /**
+//     * Test: HelperString::pregMatchAllWithOffsets
+//     */
+//    public function testPregMatchAllWithOffsets(): void
+//    {
+//        $this->assertEquals(
+//            json_encode([2 => 'a', 10 => 'a', 21 => 'a', 25 => 'a', 30 => 'a', 35 => 'a', 40 => 'a']),
+//            json_encode(HelperString::preg_match_all_with_offsets('/a/', 'Fraticinidas velum, tanquam flavum adiurator.')));
+//
+//        $this->assertEquals(
+//            json_encode([11 => 's9', 14 => 'v9', 42 => 'm9']),
+//            json_encode(HelperString::preg_match_all_with_offsets('/[a-z]9/', 'Fraticinidas9 v9 el-1 um2, tanquam 9 flavum9 adiurator39.')));
+//
+//        $this->assertEquals(
+//            json_encode([1 => 'raticinidas9', 14 => 'v9', 37 => 'flavum9']),
+//            json_encode(HelperString::preg_match_all_with_offsets('/[a-z]+9/', 'Fraticinidas9 v9 el-1 um2, tanquam 9 flavum9 adiurator39.')));
+//    }
 
     /**
      * Test: HelperString::getRandomString
@@ -463,13 +463,13 @@ class HelperStringTest extends HelperTestCase
         $this->assertTrue(HelperString::compareValuesByComparisonOperators(24, 22));
     }
 
-    /**
-     * Test: HelperString::mb_str_split
-     */
-    public function testMb_str_split(): void
-    {
-        $this->assertEquals(['t','3',' ','_','e','0','s','t'], HelperString::mb_str_split('t3 _e0st'));
-    }
+//    /**
+//     * Test: HelperString::mb_str_split
+//     */
+//    public function testMbStrSplit(): void
+//    {
+//        $this->assertEquals(['t','3',' ','_','e','0','s','t'], HelperString::mbStrSplit('t3 _e0st'));
+//    }
 
     public function testTranslate(): void
     {